+++ R/predict.R
@@ -14,9 +14,23 @@
 #'  predict function that will be used when `type = "raw"`. The
 #'  list should not include options for the model object or the
 #'  new data being predicted.
-#' @param ... Ignored. To pass arguments to pass to the underlying
-#'  function when `predict.model_fit(type = "raw")`,
-#' use the `opts` argument.
+#' @param ... Arguments to the underlying model's prediction
+#'  function cannot be passed here (see `opts`). There are some
+#'  `parsnip` related options that can be passed, depending on the
+#'  value of `type`. Possible arguments are:
+#'  \itemize{
+#'     \item `level`: for `type`s of "conf_int" and "pred_int" this
+#'            is the parameter for the tail area of the intervals
+#'            (e.g. confidence level for confidence intervals).
+#'            Default value is 0.95.
+#'     \item `std_error`: add the standard error of fit or
+#'            prediction for `type`s of "conf_int" and "pred_int".
+#'            Default value is `FALSE`.
+#'     \item `quantile`: the quantile(s) for quantile regression
+#'            (not implemented yet)
+#'     \item `time`: the time(s) for hazard probability estimates
+#'            (not implemented yet)
+#'  }
 #' @details If "type" is not supplied to `predict()`, then a choice
 #'  is made (`type = "numeric"` for regression models and
 #'  `type = "class"` for classification).
@@ -49,15 +63,18 @@
 #'  a list-column. Each list element contains a tibble with columns
 #'  `.pred` and `.quantile` (and perhaps other columns).
 #'
-#' Using `type = "raw"` with `predict.model_fit()` (or using
-#'  `predict_raw()`) will return the unadulterated results of the
-#'  prediction function.
+#' Using `type = "raw"` with `predict.model_fit()` will return
+#'  the unadulterated results of the prediction function.
 #'
 #' In the case of Spark-based models, since table columns cannot
 #'  contain dots, the same convention is used except 1) no dots
 #'  appear in names and 2) vectors are never returned but
 #'  type-specific prediction functions.
 #'
+#' When the model fit failed and the error was captured, the
+#'  `predict()` function will return the same structure as above but
+#'  filled with missing values. This does not currently work for
+#'  multivariate models.
 #' @examples
 #' library(dplyr)
 #'
@@ -90,9 +107,25 @@
 #' @method predict model_fit
 #' @export predict.model_fit
 #' @export
-predict.model_fit <- function (object, new_data, type = NULL, opts = list(), ...) {
-  if (any(names(enquos(...)) == "newdata"))
+predict.model_fit <- function(object, new_data, type = NULL, opts = list(), ...) {
+  the_dots <- enquos(...)
+  if (any(names(the_dots) == "newdata"))
     stop("Did you mean to use `new_data` instead of `newdata`?", call. = FALSE)
+
+  if (inherits(object$fit, "try-error")) {
+    warning("Model fit failed; cannot make predictions.", call. = FALSE)
+    return(NULL)
+  }
+
+  other_args <- c("level", "std_error", "quantile") # "time" for survival probs later
+  is_pred_arg <- names(the_dots) %in% other_args
+  if (any(!is_pred_arg)) {
+    bad_args <- names(the_dots)[!is_pred_arg]
+    bad_args <- paste0("`", bad_args, "`", collapse = ", ")
+    stop("The ellipses are not used to pass args to the model function's ",
+         "predict function. These arguments cannot be used: ",
+         bad_args, call. = FALSE)
+  }
 
   type <- check_pred_type(object, type)
   if (type != "raw" && length(opts) > 0)
@@ -214,14 +247,19 @@ prepare_data <- function(object, new_data) {
 #'  multiple rows per sub-model.
 #' @keywords internal
 #' @export
-multi_predict <- function(object, ...)
+multi_predict <- function(object, ...) {
+  if (inherits(object$fit, "try-error")) {
+    warning("Model fit failed; cannot make predictions.", call. = FALSE)
+    return(NULL)
+  }
   UseMethod("multi_predict")
+}
 
 #' @keywords internal
 #' @export
 #' @rdname multi_predict
 multi_predict.default <- function(object, ...)
-  stop ("No `multi_predict` method exists for objects with classes ",
+  stop("No `multi_predict` method exists for objects with classes ",
         paste0("'", class(), "'", collapse = ", "), call. = FALSE)
 
 #' @export