+++ tests/testthat/test_logistic_reg_stan.R
@@ -63,7 +63,7 @@ test_that('stan_glm prediction', {
   xy_pred <- factor(xy_pred, levels = levels(lending_club$Class))
   xy_pred <- unname(xy_pred)
 
-  expect_equal(xy_pred, predict_class(xy_fit, lending_club[1:7, num_pred]))
+  expect_equal(xy_pred, parsnip:::predict_class(xy_fit, lending_club[1:7, num_pred]))
 
   res_form <- fit(
     logistic_reg() %>%
@@ -80,7 +80,7 @@ test_that('stan_glm prediction', {
   form_pred <- unname(form_pred)
   form_pred <- ifelse(form_pred >= 0.5, "good", "bad")
   form_pred <- factor(form_pred, levels = levels(lending_club$Class))
-  expect_equal(form_pred, predict_class(res_form, lending_club[1:7, c("funded_amnt", "int_rate")]))
+  expect_equal(form_pred, parsnip:::predict_class(res_form, lending_club[1:7, c("funded_amnt", "int_rate")]))
 })
 
 
@@ -102,7 +102,7 @@ test_that('stan_glm probability', {
   xy_pred <- xy_fit$fit$family$linkinv(xy_pred)
   xy_pred <- tibble(bad = 1 - xy_pred, good = xy_pred)
 
-  expect_equal(xy_pred, predict_classprob(xy_fit, lending_club[1:7, num_pred]))
+  expect_equal(xy_pred, parsnip:::predict_classprob(xy_fit, lending_club[1:7, num_pred]))
 
   res_form <- fit(
     logistic_reg() %>%
@@ -117,7 +117,7 @@ test_that('stan_glm probability', {
             newdata = lending_club[1:7, c("funded_amnt", "int_rate")])
   form_pred <- xy_fit$fit$family$linkinv(form_pred)
   form_pred <- tibble(bad = 1 - form_pred, good = form_pred)
-  expect_equal(form_pred, predict_classprob(res_form, lending_club[1:7, c("funded_amnt", "int_rate")]))
+  expect_equal(form_pred, parsnip:::predict_classprob(res_form, lending_club[1:7, c("funded_amnt", "int_rate")]))
 })
 
 
@@ -156,8 +156,10 @@ test_that('stan intervals', {
   stan_upper <- apply(stan_post, 2, quantile, prob = 0.965)
   stan_std  <- apply(stan_post, 2, sd)
 
-  expect_equivalent(confidence_parsnip$.pred_lower, stan_lower)
-  expect_equivalent(confidence_parsnip$.pred_upper, stan_upper)
+  expect_equivalent(confidence_parsnip$.pred_lower_good, stan_lower)
+  expect_equivalent(confidence_parsnip$.pred_upper_good, stan_upper)
+  expect_equivalent(confidence_parsnip$.pred_lower_bad, 1 - stan_upper)
+  expect_equivalent(confidence_parsnip$.pred_upper_bad, 1 - stan_lower)
   expect_equivalent(confidence_parsnip$.std_error, stan_std)
 
   stan_pred_post <-
@@ -168,8 +170,10 @@ test_that('stan intervals', {
   stan_pred_upper <- apply(stan_pred_post, 2, quantile, prob = 0.965)
   stan_pred_std  <- apply(stan_pred_post, 2, sd)
 
-  expect_equivalent(prediction_parsnip$.pred_lower, stan_pred_lower)
-  expect_equivalent(prediction_parsnip$.pred_upper, stan_pred_upper)
+  expect_equivalent(prediction_parsnip$.pred_lower_good, stan_pred_lower)
+  expect_equivalent(prediction_parsnip$.pred_upper_good, stan_pred_upper)
+  expect_equivalent(prediction_parsnip$.pred_lower_bad, 1 - stan_pred_upper)
+  expect_equivalent(prediction_parsnip$.pred_upper_bad, 1 - stan_pred_lower)
   expect_equivalent(prediction_parsnip$.std_error, stan_pred_std, tolerance = 0.1)
 })