+++ tests/testthat/test_multinom_reg_spark.R
@@ -45,7 +45,7 @@ test_that('spark execution', {
   )
 
   expect_error(
-    spark_class_pred_class <- predict_class(spark_class_fit, iris_te),
+    spark_class_pred_class <- parsnip:::predict_class(spark_class_fit, iris_te),
     regexp = NA
   )
 
@@ -62,7 +62,7 @@ test_that('spark execution', {
   )
 
   expect_error(
-    spark_class_prob_classprob <- predict_classprob(spark_class_fit, iris_te),
+    spark_class_prob_classprob <- parsnip:::predict_classprob(spark_class_fit, iris_te),
     regexp = NA
   )